+++ Chat_App/Chat_App/Constants/AppColors.swift
@@ -5,21 +5,21 @@
 //  Created by Gi Oo on 07.05.23.
 //
 
-import UIKit
-
-enum AppColors {
-    static let dividerViewColor = UIColor(red: 0.97,
-                                          green: 0.81,
-                                          blue: 0.50,
-                                          alpha: 1.0)
-
-    static let darkModeColor = UIColor(red: 22/255,
-                                       green: 0,
-                                       blue: 57/255,
-                                       alpha: 1)
-  
-    static let containerBorderColor = UIColor(red: 0.62,
-                                             green: 0.38,
-                                             blue: 1.00,
-                                             alpha: 1.00)
-}
+//import UIKit
+//
+//enum AppColors {
+//    static let dividerViewColor = UIColor(red: 0.97,
+//                                          green: 0.81,
+//                                          blue: 0.50,
+//                                          alpha: 1.0)
+//
+//    static let darkModeColor = UIColor(red: 22/255,
+//                                       green: 0,
+//                                       blue: 57/255,
+//                                       alpha: 1)
+//  
+//    static let containerBorderColor = UIColor(red: 0.62,
+//                                             green: 0.38,
+//                                             blue: 1.00,
+//                                             alpha: 1.00)
+//}
