+++ Chat_App/switcher.swift
@@ -1,0 +1,12 @@
+//
+//  switcher.swift
+//  Chat_App
+//
+//  Created by Gi Oo on 08.05.23.
+//
+
+import UIKit
+
+class SwitchButton: UIButton {
+
+}
